Add Redis caching supportfor external API calls
Each page load would send a bunch of requests to the competition service, which
could have been a problem if more than a few people try to use the site. The
JSON data is now cached in Redis and is set to expire after five minutes. If
someone tries to view the site after five minutes has taken place, the app will
request more up to date information from the source.