Return request promise when cache misses
This caused a bug where the app would fail when a new competition was added to
the event list, but the event wasn't cached in Redis. It wasn't too terrible of
a problem, since it would still be added to cache and would cause a hit on
reload, but it was still annoying to have it break for the first load after new
scores had been posted.